API

API documentation

Index

Mortar3D.calculate_normals

Mortar3D.project_from_plane_to_surface

Mortar3D.project_from_surface_to_plane

Mortar3D.calculate_normals — Method.

calculate_normals(elements::Dict{Int, Vector{Int}},
                  element_types::Dict{Int, Symbol},
                  X::Dict{Int, Vector{Float64})

Given elements, element types and node locations, calculate nodal normals by first calculating normal directions for each element and then averaging them in nodes. As a result we get unique normal direction defined to each node.

Notes

Only linear elements supported.

Example

X = Dict(1 => [0.0, 0.0, 0.0], 2 => [1.0, 0.0, 0.0], 3 => [0.0, 1.0, 0.0])
elements = Dict(1 => [1, 2, 3])
element_types = Dict(1 => :Tri3)
normals = calculate_normals(elements, element_types, X)

# output

Dict{Int64,Array{Float64,1}} with 3 entries:
  2 => [0.0, 0.0, 1.0]
  3 => [0.0, 0.0, 1.0]
  1 => [0.0, 0.0, 1.0]

source

Mortar3D.project_from_plane_to_surface — Method.

project_from_plane_to_surface(Val{:Tri3}, p, n, x1, x2, x3)

Project vertex p to surface of Tri3 element.

Example

p = [1/3, 1/3, 0.0]
n = [0.0, 0.0, 1.0]
x1 = [0.0, 0.0, 1.0]
x2 = [1.0, 0.0, 1.0]
x3 = [0.0, 1.0, 1.0]
theta, d = project_from_plane_to_surface(Val{:Tri3}, p, n, x1, x2, x3)
round.(theta, 3), d

# output

([0.333, 0.333], 1.0)

source

Mortar3D.project_from_surface_to_plane — Method.

project_from_surface_to_plane(p, x0, n0)

Find the projection of vertex p to the auxiliary plane (x0, n0), where x0 is the origo of plane and n0 is unit normal of plane.

Example

p = [1.0, 1.0, 3.0]
x0 = [0.0, 0.0, 0.0]
n0 = [0.0, 0.0, 1.0]
project_from_surface_to_plane(p, x0, n0)

# output
3-element Array{Float64,1}:
 1.0
 1.0
 0.0

source