API

API documentation

Index

Mortar3D.approx_in

Mortar3D.calculate_normals

Mortar3D.calculate_polygon_clip

Mortar3D.check_orientation!

Mortar3D.is_vertex_inside_polygon

Mortar3D.project_from_plane_to_surface

Mortar3D.project_from_surface_to_plane

Mortar3D.approx_in — Method.

approx_in(q, P)

Test does vector P contain approximately q. This function uses isapprox() internally to make boolean test.

Example

P = Vector[[1.0, 1.0], [2.0, 2.0]]
q = [1.0, 1.0] + eps(Float64)
in(q, P), approx_in(q, P)

# output

(false, true)

source

Mortar3D.calculate_normals — Method.

calculate_normals(elements::Dict{Int, Vector{Int}},
                  element_types::Dict{Int, Symbol},
                  X::Dict{Int, Vector{Float64})

Given elements, element types and node locations, calculate nodal normals by first calculating normal directions for each element and then averaging them in nodes. As a result we get unique normal direction defined to each node.

Notes

Only linear elements supported.

Example

X = Dict(1 => [0.0, 0.0, 0.0], 2 => [1.0, 0.0, 0.0], 3 => [0.0, 1.0, 0.0])
elements = Dict(1 => [1, 2, 3])
element_types = Dict(1 => :Tri3)
normals = calculate_normals(elements, element_types, X)

# output

Dict{Int64,Array{Float64,1}} with 3 entries:
  2 => [0.0, 0.0, 1.0]
  3 => [0.0, 0.0, 1.0]
  1 => [0.0, 0.0, 1.0]

source

Mortar3D.calculate_polygon_clip — Method.

calculate_polygon_clip(xs, xm, n)

Given two polygons xs and xm, calculate polygon clipping P. It is assumed that clipping is done in direction n.

Example

xs = Vector[[0, 0, 0], [1, 0, 0], [0, 1, 0]]
xm = Vector[[-1/4, 1/2, 0], [1/2, -1/4, 0], [3/4, 3/4, 0]]
n = [0, 0, 1]
P = calculate_polygon_clip(xs, xm, n)

# output

6-element Array{Array{T,1} where T,1}:
 [0.65, 0.35, 0.0]
 [0.5625, 0.0, 0.0]
 [0.25, 0.0, 0.0]
 [0.0, 0.25, 0.0]
 [0.0, 0.5625, 0.0]
 [0.35, 0.65, 0.0]

source

Mortar3D.check_orientation! — Method.

check_orientation!(P, n)

Given polygon P and normal direction n, ensure that polygon vertices are ordered in counter clock wise direction with respect to surface normal. It is assumed that polygon is convex.

Example

Unit triangle, normal in z-direction:

P = Vector[[0.0, 0.0, 0.0], [0.0, 1.0, 0.0], [1.0, 0.0, 0.0]]
n = [0.0, 0.0, 1.0]
check_orientation!(P, n)

# output

3-element Array{Array{T,1} where T,1}:
 [1.0, 0.0, 0.0]
 [0.0, 0.0, 0.0]
 [0.0, 1.0, 0.0]

source

Mortar3D.is_vertex_inside_polygon — Method.

is_vertex_inside_polygon(q, P; atol=1.0e-3)

Test is vertex q inside polygon P with given tolerance.

Example

P = Vector[[0.0,0.0,0.0], [1.0,0.0,0.0], [0.0,1.0,0.0]]
q1 = [ 0.1, 0.0, 0.0]
q2 = [-0.1, 0.1, 0.0]
is_vertex_inside_polygon(q1, P), is_vertex_inside_polygon(q2, P)

# output

(true, false)

source

Mortar3D.project_from_plane_to_surface — Method.

project_from_plane_to_surface(Val{:Tri3}, p, n, x1, x2, x3)

Project vertex p to surface of Tri3 element.

Example

p = [1/3, 1/3, 0.0]
n = [0.0, 0.0, 1.0]
x1 = [0.0, 0.0, 1.0]
x2 = [1.0, 0.0, 1.0]
x3 = [0.0, 1.0, 1.0]
theta, d = project_from_plane_to_surface(Val{:Tri3}, p, n, x1, x2, x3)
round.(theta, 3), d

# output

([0.333, 0.333], 1.0)

source

Mortar3D.project_from_surface_to_plane — Method.

project_from_surface_to_plane(p, x0, n0)

Find the projection of vertex p to the auxiliary plane (x0, n0), where x0 is the origo of plane and n0 is unit normal of plane.

Example

p = [1.0, 1.0, 3.0]
x0 = [0.0, 0.0, 0.0]
n0 = [0.0, 0.0, 1.0]
project_from_surface_to_plane(p, x0, n0)

# output
3-element Array{Float64,1}:
 1.0
 1.0
 0.0

source